Update API endpoints and add user login session

+++ public/assets/js/my_posts.js
@@ -1,27 +1,30 @@
 // get table tbody element from html
 const tbody = document.querySelector( 'tbody' );
+const user_id = sessionStorage.getItem( 'user_id' );
 
 // get all posts from the database
-fetch( 'http://127.0.0.1:3000/posts' )
+fetch( `http://127.0.0.1:8080/posts` )
     .then( res => res.json() )
     .then( data => {
         // loop through the posts and add them to the table
         data.forEach( post => {
-            // format date created
-            const fullDate = new Date( post.created_at );
-            const date = fullDate.toLocaleDateString();
+            if ( post.author_id == user_id ) {
+                // format date created
+                const fullDate = new Date( post.created_at );
+                const date = fullDate.toLocaleDateString();
 
-            tbody.innerHTML += `
+                tbody.innerHTML += `
                 <tr>
-                    <td>${post.id}</td>
-                    <td>${post.post_title}</td>
-                    <td>${date}</td>
+                    <td>${ post.id }</td>
+                    <td>${ post.post_title }</td>
+                    <td>${ date }</td>
                     <td>
-                        <a href="./create-post.html?postID=${post.id}">Edit</a>
-                        <a href="./delete-post.html?postID=${post.id}">Delete</a>
+                        <a href="./create-post.html?postID=${ post.id }">Edit</a>
+                        <a href="./delete-post.html?postID=${ post.id }">Delete</a>
                     </td>  
                 </tr>
             `;
+            }
         } );
     } )
     .catch( err => console.log( err ) );
